refactor(translator): split translator into files based on FileType
